Add dvcoapbr to iset.mm
This is dvcobr from set.mm but with a number of changes related to
whether the function G is well behaved enough.  This replaces the if
expressions which were built on real number trichotomy in set.mm.
Other than that, though, the theorem is much as stated in set.mm and the
proof is similar to the set.mm proof.