Add dedekindeu to iset.mm
This is a bit like dedekind in set.mm although various details are
different including that the lower and upper cuts have to be open (and
thus the real corresponding to the Dedekind cut is not contained
in either one).

Includes lemmas dedekindeulemuub , dedekindeulemub , dedekindeulemloc ,
dedekindeulemlub , dedekindeulemlu , and dedekindeulemeu

Expand dedekind entry in mmil.html to mention dedekindeu and
differences.