Add suplocsr to iset.mm
This is supsr from set.mm but for a located set and with many parts of
the proof needing intuitionizing (no big conceptual changes other than
the located part, just a lot of details).

Includes lemma suplocsrlem which is similar to supsrlem in set.mm
although most parts of the proof need intuitionizing or rewriting.

Includes lemmas suplocsrlemb and suplocsrlempr

Add mappsrprg . This is mappsrpr from set.mm but one direction of the
biconditional and in closed form.

Add map2psrprg .  This is map2psrpr from set.mm but in closed form.

Add ltpsrprg .  This is ltpsrpr from set.mm with additional set
existence conditions.  The proof is similar to the set.mm proof.