Add sinhalfpilem to iset.mm
Stated as in set.mm.  The proof requires a fair bit of intuitionizing
(it was using excluded middle for things which didn't need it, like
dividing by a positive real), but generally follows the set.mm proof.

Mention sq11 in the mmil.html entry for sqeqori , because sq11 is a
special case which iset.mm does have.