Add several sin/cos theorems to iset.mm
sinq12gt0 is stated as in set.mm and the proof is the set.mm proof with
very minor changes related to not equal versus apart.

sinq34lt0t and cosq14gt0 are copied without change from set.mm

cosq23lt0 is a new theorem which is inspired by cosq14gt0 with a
somewhat similar proof.

Add coseq0q4123 - this is a new theorem similar to coseq00topi from
set.mm but showing that pi/2 is the only zero for cosine over a wider
range.  The proof is significantly different from the set.mm proof of
coseq00topi .

coseq00topi is stated as in set.mm but the proof here is a simple
consequence of coseq0q4123 .

The theorems are reordered from set.mm because the coseq00topi proof
uses the other theorems added here.