Add exmidonfin to iset.mm (#3861)
We are unable to prove that a finite ordinal is a natural number.  The
proof is a formalization of an informal proof by Andrew W Swan posted on
mathstodon.xyz.

Includes lemma exmidonfinlem

Update mmil.html because there are some theorems we know are not possible
and others that we suspect are not.