Add coseq0negpitopi to iset.mm
Stated as in set.mm.

The proof is somewhat the same as the set.mm proof, but the big
difference is that instead of dividing the range for A into no greater
than zero and no less than zero, we need two overlapping ranges.
Fortunately, that isn't especially difficult to arrange.