Link to theorems in theorem sampler
The main reason to link using the ~ syntax is that the build process
will verify that these theorems exist.  Showing the names to the user is
maybe good, maybe bad, but is mainly here as a side effect of that
verification.

Rename prth to anim12 as this theorem was renamed a while ago.