Janitorial work (#3891)
* elabd renamed to spcedv

The label "elabd" seemed to be a mistake: the theorem is not about class abstraction, but about existential specialisation. Therefore, the label was changed to "spcedv", the theorem was moved up, and the comment was adjusted.

* "real" elabd added

The "real" theorem ~elabd (about class abstraction) was added. Proofs of theorems  ~sursubmefmnd, ~injsubmefmnd, ~wfrlem15 could have been shortened,

* anim12da   duplicate of anim12dan

~anim12da deleted in JM's mathbox (duplicate of ~anim12dan)

I have changed the contributor of ~anim12dan to JF, because ~anim12da was created earlier.

* proofs shortened by using uniexd

* Rewrap

* hypothesis of elabd modified

`elabd.1 $e |- ( ph -> A e. _V ) $.` replaced by `elabd.1 $e |- ( ph -> A e. V ) $.` (according to our conventions).
7 additional proof could have been shortened afterwards.