Add prodmodclem3 and prodmodclem2a to iset.mm
prodmodclem3 is prodmolem3 with some modifications for differences between
iset.mm and set.mm.  The proof is based on the set.mm proof with
significant modifications in various parts to deal with differences in
seq theorems and the like.

prodmodclem2a is prodmolem2a from set.mm with various changes to the
statement of the theorem and the proof.  The general idea of the lemma
and proof remain the same, however.