Update discussion of ruc in mmcomplex.html (#3931)
The current text has diverged quite a bit from the proof in set.mm (for
example, it refers to 39 lemmas, but set.mm has 13).  Update it to
reflect the way the set.mm proof now works, but keep the overall
structure, and most of the text, intact.

Refer to theorems (including lemmas) by using the ~ syntax which should
reduce the chance of this happening again if the set.mm proof is changed
again to this extent.