Add ioocosf1o to iset.mm
This is recosf1o from set.mm but for an open interval rather than a
closed interval.  The closed interval theorem might be possible too but the
open interval theorem is what we prove here.  The proof is similar to
the set.mm proof with differences such as the lack of ivthle2 .