Add cos0pilt1 to set.mm and iset.mm
This follows relatively directly from various cosine theorems we already
have.  The immediate need is in iset.mm but I believe our preference is
to keep such theorems both in set.mm and iset.mm where it easily can be done.

The proof is the same in set.mm and iset.mm.