Intuitionize trigonometry to the end of the Properties of pi section (#3980)
* Add cos11 to iset.mm

Stated as in set.mm.  The proof is from cosordlem rather than from
cosord and is different in various details, but is somewhat similar to
the set.mm proof of cos11 .

Add cosord to mmil.html.

* Add ivthle2 and sinord to mmil.html

* Add cos0pilt1 to set.mm and iset.mm

This follows relatively directly from various cosine theorems we already
have.  The immediate need is in iset.mm but I believe our preference is
to keep such theorems both in set.mm and iset.mm where it easily can be done.

The proof is the same in set.mm and iset.mm.

* Mark rgen2a as discouraged in iset.mm.

It is already discouraged in set.mm and basically the same
considerations apply to iset.mm too.

* Add ioocosf1o to iset.mm

This is recosf1o from set.mm but for an open interval rather than a
closed interval.  The closed interval theorem might be possible too but the
open interval theorem is what we prove here.  The proof is similar to
the set.mm proof with differences such as the lack of ivthle2 .

* Add resinf1o to mmil.html

* add tanord1 and tanord to mmil.html

* add tanregt0 to mmil.html

* Add negpitopissre to iset.mm

Copied without change from set.mm.