Replace syl5rbbr with bitr3di
This is in set.mm and iset.mm

bitr3di is syl5rbbr with the order of the hypotheses switched.

Run "minimize_with bitr3di/may_grow" on each of the theorems which had
been using syl5rbbr .

For theorems marked as "Proof modification is discouraged", use
/override on PROVE and SAVE NEW_PROOF, but check that the change from
syl5rbbr to bitr3di is the only modification printed in the output.