Add reeff1oleme to iset.mm
Similar to reeff1olem but over a different portion of the real line.
The proof is loosely based on the set.mm proof of reeff1o (at least in
the sense of looking at the reciprocal) but unlike in set.mm we don't
just need intervals which abut at one, we need intervals which overlap.