Intuitionize (real) natural logarithm from df-log through relogdiv (#3992)
* Add df-relog to iset.mm

This is based on dfrelog in set.mm but we use it as our definition for
logarithm, because defining the complex logarithm seems like it won't
work (at least not as in set.mm).

Mention in mmil.html.

* add df-cxp to mmil.html

* Add logrn , ellogrn , dflog2 to mmil.html

* Add reeff1olem to iset.mm

Stated as in set.mm.  The proof is very similar to the set.mm proof but
modified to use ivthinc in place of ivth .

* Add reeff1oleme to iset.mm

Similar to reeff1olem but over a different portion of the real line.
The proof is loosely based on the set.mm proof of reeff1o (at least in
the sense of looking at the reciprocal) but unlike in set.mm we don't
just need intervals which abut at one, we need intervals which overlap.

* Add reeff1o to set.mm

Stated as in set.mm.  The proof is similar to the set.mm proof except
that instead of having two intervals which abut at one, it needs two
intervals which overlap.

* add relogrn to mmil.html

* add logrncn , eff1o2 , logf1o to mmil.html

* Add dfrelog to iset.mm

Stated as in set.mm.  The proof is a fairly direct consequence of
reeff1o and df-relog .

* Add theorems to mmil.html

These are the theorems in the "Mapping of the exponential function"
section from set.mm and none of them intuitionize easily.

They are efgh , efif1o , efifo , eff1o , efabl , efsubm ,
circgrp , and circsubm

* Add relogf1o to iset.mm

Stated as in set.mm.  The proof is direct from reeff1o and dfrelog .

* Add complex logarithm theorems to mmil.html

This is logrncl , logcl , logimcl , logcld , logimcld , logimclad ,
abslogimle , and logrnaddcl

* Add relogcl to iset.mm

Copied without change from set.mm.

* add eflog , logeq0im1 , logccne0 to mmil.html

* add logne0 to mmil.html

* Add reeflog to iset.mm

Stated as in set.mm.  The proof is a new one but is not difficult.

* Intuitionize logef and relogef

Add logef to mmil.html

Add reflogef to iset.mm.  Stated as in set.mm.  It is proved as a
consequence of reeflog and reeff1 .

* add logeftb to mmil.html

* Add relogeftb to iset.mm

Stated as in set.mm.  The proof is from reeflog and relogef .

* Add log1 and loge to iset.mm

Copied without change from set.mm.

* add logneg , logm1 , lognegb to mmil.html

* Add relogoprlem , relogmul and relogdiv to iset.mm

Copied without change from set.mm.