Add pw1nct to iset.mm
This is an exercise.  The full motivation is hard to describe in iset.mm
terms but from what I gather from Mastodon, a realizability topos
satisfies the uniformity principle shown here and ~P 1o is as usual a
representation of truth values (a subobject classifier, or maybe it is
better to call it isomorphic to a subobject classifier).