Add prodssdc to iset.mm
This is prodss from set.mm with two decidability hypotheses added plus a
hypothesis for M e. ZZ to replace the use of case elimination on M e. ZZ
in the set.mm proof.  The proof is basically the set.mm proof but most
steps need some kind of intuitionizing.  See also isumss in iset.mm for
a similar theorem for sums.