Add fprod2d to iset.mm
Stated as in set.mm.  The proof is the same as the set.mm proof except
for the need for pass one more hypothesis to fprod2dlemstep which is
basically already present and just needs to have a little logic
rearranged.

Includes fprod2dlemstep which is fprod2dlem from set.mm with one
more hypothesis.  The proof needs intuitionizing in places but is
basically the set.mm proof.  In some places it is inspired by the
iset.mm proof of fsum2dlemstep .