Add cbvalv1 and cbvexv1 to iset.mm
The proof of cbvalv1 is taken from set.mm.  The proof of cbvexv1 is
based on the iset.mm proof of cbvexh .

Also, add axc4i , equsalv , nfd2 (this is what is nfd in set.mm),
hbe1a , nf5-1 , nf5d , nfabdw , nfsbcdw , and nfcsbw to iset.mm.
Most of these are stated as in set.mm with either the same proof or a
similar proof as set.mm.