Add fprodsplit1f to iset.mm
Stated as in set.mm.  The proof needs a bit of intuitionizing but is
mostly the same as the set.mm proof.  Modify the comment to not refer to
a theorem which is in a mathbox in set.mm and which I therefore don't
plan on copying to iset.mm at this time.