remove axioms in copsex2g; revise opelopabsbALT
justification comments:

the copsex2g change was mentioned in #4176

uncommute opelopabsbALT, it is more natural this way and shortens proofs

opelopabsbALT can now be auto-used in opelopabsb; in my opinion it is different enough to be its own theorem rather than an alternate, hence the rename. I don't see a reason to discourage its usage.

acos1half is simply a random unused mathbox theorem