Intuitionize finite products from fprodxp to fprodsplitsn (#4175)
* Add fprodxp to iset.mm

Copied without change from set.mm.

* Add fprodcnv to iset.mm

Stated as in set.mm.  The proof is similar to the set.mm proof but needs
intuitionizing in a number of places.

* Add fprodcom2fi to iset.mm

This is fprodcom2 from set.mm with one additional hypothesis.  The proof
is based on the set.mm proof of fprodcom2 and the iset.mm proof of
fisumcom2 .

* Add fprodcom to iset.mm

Stated as in set.mm.  The proof is similar to the set.mm proof.

* Add fprod0diagfz to iset.mm

This is fprod0diag from set.mm with a N e. ZZ condition added.  The
proof is similar to the set.mm proof.

* Add fproddivap to iset.mm

This is fproddiv from set.mm with not equal changed to apart.  The proof
is in terms of fprodmul and a new theorem fprodrec .

Includes fprodrec which is similar but where the numerator is one.
The proof is by finite set induction.

* Add spimfv and chvarfv to iset.mm

Copied from set.mm except:

* adjustments to the comments to reflect differences between iset.mm and
  set.mm
* the theorem which is ax6ev in set.mm is a9ev in iset.mm.  Also note
  this in mmil.html.

* Add fproddivapf to iset.mm

This is fproddivf from set.mm with not equal changed to apart.  The
proof is similar to the set.mm proof.

* Add fprodsplitf to iset.mm

Copied without change from set.mm.

* Add fprodsplitsn to iset.mm

Stated as in set.mm.  The logic in the proof needs some rearrangement to
be intuitionistic, but the proof is basically the set.mm proof.