Wording tweak to ramsey comment in set.mm
Without this change "with the definition Ramsey eliminated" makes it sound
like Ramsey (the person) stated this theorem without a definition, and
we are putting it back or something.  At least, that's how I read it at
first (the typography does not show the backticks clearly, at least
using my browser).