Add pythagtrip to iset.mm
Stated as in set.mm.  The proof needs a bit of intuitionizing but is
basically the set.mm proof.

Specifically, lemmas that are copied without change are pythagtriplem1 ,
pythagtriplem3 , pythagtriplem4 , pythagtriplem6 , pythagtriplem7 ,
pythagtriplem8 , pythagtriplem9 , pythagtriplem11 , pythagtriplem13 ,
and pythagtriplem18 .

Lemmas that are stated as in set.mm but which need some intuitionizing
in the proof (but basically can use the set.mm proof) are: pythagtriplem2 ,
pythagtriplem10 , pythagtriplem12 , pythagtriplem14 , pythagtriplem15 ,
pythagtriplem16 , pythagtriplem17 , pythagtriplem19 , and pythagtrip itself.

Update mm_100.html and mmil.html