Replace syl5reqr with eqtr3di
This is in set.mm and iset.mm

* Add eqtr3di with the order of the hypotheses switched and the variables renamed.

* edit local copy of scripts/minimize to use "minimize_with eqtr3di/may_grow" in
place of the normal minimize

* Run "show usage syl5reqr" and copy-paste the output to a shell command like
for i in x y z; do scripts/minimize $i set.mm; done

* For theorems which had been marked "Proof modification is discouraged",
add /OVERRIDE on PROVE and SAVE NEW_PROOF in the script and run.
Check that the change to use eqtr3di is the only modification printed in the
output. (This step was needed for sineq0ALT in set.mm and nothing in iset.mm)

* Remove syl5reqr

* Repeat for iset.mm