Intuitionize pCnt through closure (#4275)
* Add pCnt to iset.mm

Copied without change from set.mm.  This is the syntax and the
definition df-pc .

* Add suprzubdc to iset.mm

This is suprzub from set.mm with an additional decidability condition.
There are probably various way to prove it, but the proof from zsupcl is
fairly simple so that's the proof here.

* Add zsupssdc to iset.mm

This is zsupss from set.mm with nonempty changed to inhabited and a
decidability condition added.  The proof is based on the set.mm proof
but is different in various ways.

Add $j usage for zsupss in set.mm to match the claim which is already in
the comment.

* Add suprzcl2dc to iset.mm

This is suprzcl2 from set.mm with a decidability condition added and
nonempty changed to inhabited.  The proof is loosely based on the set.mm
proof.

Add $j usage to set.mm to match what was already in the suprzcl2
comment.

* Add ssrab3 to iset.mm

Copied without change from set.mm

* Add ltexp2 to iset.mm

Stated as in set.mm.  This proof is via ^c .

* Add nn0ltexp2 to iset.mm

Turns out that ltexp2 for NN0 is not especially hard to prove by
induction.

* Add elnn0dc to iset.mm

This is a pretty straightforward variation of eluzdc .

* Intuitionize pclem

Rather than have the conclusion be several conjuncts, split the theorem
into several theorems pclem0 , pclemub , and pclemdc

* Add pcprecl to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
similar to the set.mm proof.

* Add pcprendvds to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is similar
to the set.mm proof.

* Add pcprendvds2 to iset.mm

Stated as in set.mm.  The proof needs a very small amount of
intuitionizing but is otherwise the set.mm proof.

* Add nn0leexp2 to iset.mm

This is leexp2 from set.mm, but for NN0 .  The proof is from nn0ltexp2 .

* Add pcpre1 to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* Add eqlelt and eqleltd to iset.mm

Copied without change from set.mm.

* Add dvdstrd to iset.mm

Copied without change from set.mm.

* Add pcpremul to iset.mm

Stated as in set.mm.  The proof needs intutionizing in various places
but is basically the set.mm proof.

* Add divmuleqapd to iset.mm

This is divmuleqd from set.mm with not equal changed to apart.  It is a
deduction form of divmuleqap which had already been in iset.mm.  The
proof is trivial.

* Add pceu and pcval to iset.mm

Stated as in set.mm.  The proofs need some intuitionizing (small
in some places, larger in others) but are basically the set.mm proofs.

Includes lemma pceulem .

* Add pczpre and pczcl to iset.mm

Stated as in set.mm.  The proof of pczpre needs a little bit of
intuitionizing (but is similar to the set.mm proof); the proof
of pczcl is unchanged from the set.mm proof.

* Add pccl and pccld to iset.mm

Copied without change from set.mm.