Add infpn2 and prminf to iset.mm
prminf is stated as in set.mm.  The proof is similar to the set.mm proof but uses
unbendc in place of unben .

infpn2 is stated as in set.mm.  The proof is in terms of prminf because
applying unbendc directly (the way that set.mm applies unben ) would
mean a more involved decidability condition and it is easier to just
take advantage of the decidability already shown as part of prminf .