Start rename of sseldi to sselid
Because there are 1010 usages in set.mm, we'll rename this in several
stages.  So introduce the new name, and rename some of the usages.

We do not use the discouraged file as that seems like it just invites
merge conflicts on the discouraged file with no particularly compelling
benefit in this situation.