Add 1arith and lemmas to iset.mm
All are stated as in set.mm.

The proofs are similar to the set.mm proofs. Some are copied without
change, some need a small amount of intuitionizing, and some need
intuitionizing more places.  But none of them need large changes to the
structure of the proof.

Lemmas are 1arithlem1 , 1arithlem2 , 1arithlem3 , and 1arithlem4