Revise Metamath 100 entry 7 in mmil.html
This is the Law of Quadratic Reciprocity and the text is out of date now
that iset.mm has df-lgs and lgs related theorems.

I took a look over the set.mm proof and didn't see a lot of obvious
other prerequisites but the proof is long enough that I could have
easily missed something.