Say Abelian group is comment for ablpncan3
This is in set.mm and iset.mm

Since those are synonymous (both are, for example, mentioned in the
opening sentence of https://en.wikipedia.org/wiki/Abelian_group ),
I'm a bit reluctant to police which one is used.

But I guess we mostly use Abelian group so :shrug: I guess I can
change this.