Define tight apartness predicate in iset.mm
TAp is a predicate saying that a relation is a tight apartness on a
class. Includes syntax and df-tap .

Add tapeq1 .  This is an equality theorem for tight apartness, analogous
to (for example) ereq1 .

Add netap to iset.mm.  Negated equality is a tight apartness if
equality is decidable.

Add 2onetap to iset.mm.  This is a special case of netap which we expect
to be useful.

Add 2oneel which says that zero does not equal one (expressed in terms
of an apartness).

Add 2omotap which shows that if there is at most one tight apartness on
2o , exluded middle follows.  Includes lemmas 2omotaplemap and 2omotaplemst .