Move dvrdir and rdivmuldivd from mathbox to main
This is in set.mm.  The immediate motivation is to use dvrdir in iset.mm
(combined with a desire to have theorems in iset.mm main also in set.mm
main).  But these do seem like generally applicable theorems for which
we already have the complex number analogues in main.