Add CCfld to iset.mm
This is the syntax and df-icnfld .  df-icnfld is based on df-cnfld from
set.mm but with only the first 4 slots (at least for now), because
showing existence of the sets for the other slots depends on
theorems we haven't proved yet at this point in the file.

Add cnfldstr to iset.mm . Stated as in set.mm.  The proof is based on a
portion of the set.mm proof.

Update iset-discouraged file.