Add ercpbl to iset.mm
Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

The distinct variable conditions here could be increased or decreased
(depending on how much variable changing we want to do), but just use
the conditions used in set.mm, for convenience moving proofs back and
forth.