Add ptex to iset.mm
This is a set existence theorem which would be covered by fvex in
set.mm.  The proof is fairly involved, because we need to find a set
which contains all the values which can be in its value, which is
defined by an unrestricted class abstraction which only indirectly
allows us to find such a containing set.

Also, move tgval and tgvalex earlier in iset.mm , after df-topgen ,
because we need tgvalex in the ptex proof.