Intuitionize quotient ring from df-qus to xpsval (#4724)
* Add /s to iset.mm

This is the syntax and df-qus .  Copied without change from set.mm.
Includes the distinct variable constraints from set.mm.

Adjust comment to refer to df-iimas .  Remove df-qus from mmil.html.

* Add qusval to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy quslem from set.mm to iset.mm

* copy qusin from set.mm to iset.mm

* copy qusbas from set.mm to iset.mm

* Add quss to mmil.html

Also some image structure theorems which seem to have been accidentally
overlooked before: imassca , imasvsca , imasip , imastset , and imasle

* Add divsfvalg to iset.mm

This is divsfval from set.mm with an A e. V condition added.  The set.mm
proof does case elimination on this condition.  The iset.mm proof is
taken from a portion of the set.mm proof.

* Add ercpbllemg to iset.mm

This is ercpbllem from set.mm with a condition added.  The proof is
similar to the set.mm proof.

* Add ercpbl to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

The distinct variable conditions here could be increased or decreased
(depending on how much variable changing we want to do), but just use
the conditions used in set.mm, for convenience moving proofs back and
forth.

* Add erlecpbl to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* Add qusaddvallemg to iset.mm

This is qusaddvallem from set.mm with a set existence condition added.
The proof needs some intuitionizing but is basically the set.mm proof.

* Add qusaddflemg to iset.mm

This is qusaddflem from set.mm with an added set existence condition.
The proof needs some intuitionizing but is basically the set.mm proof.

* Add qusaddval to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.

* Add qusaddf to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.

* Add qusmulval to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* Add qusmulf to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* copy fnpr2o from set.mm to iset.mm

* copy fnpr2ob from set.mm to iset.mm

* copy fvpr0o from set.mm to iset.mm

* copy fvpr1o from set.mm to iset.mm

* copy fvprif and xpsfrnel from set.mm to iset.mm

* Add xpsfeq to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy xpsfrnel2 from set.mm to iset.mm

* Add xpscf to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add xpsfval to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add xpsff1o to iset.mm

Stated as in set.mm.  The proof needs a small amount of inuitionizing
but is basically the set.mm proof.

* copy xpsfrn from set.mm to iset.mm

* copy xpsff1o2 from set.mm to iset.mm

* Add binary structure product to iset.mm

This is the syntax and df-xps .  Copied without change.

* Add ptex to iset.mm

This is a set existence theorem which would be covered by fvex in
set.mm.  The proof is fairly involved, because we need to find a set
which contains all the values which can be in its value, which is
defined by an unrestricted class abstraction which only indirectly
allows us to find such a containing set.

Also, move tgval and tgvalex earlier in iset.mm , after df-topgen ,
because we need tgvalex in the ptex proof.

* copy homid from set.mm to iset.mm

* Add homslid to iset.mm

The usual slot theorem.

* Add ccoid and ccoslid to iset.mm

ccoid is copied without change from set.mm.

ccoslid is one of the usual slot theorems.

* Add prdsex to iset.mm

This is a set existence theorem, which would be covered by ovex in
set.mm.  The proof is fairly straightforward but long as there are quite
a few slots covered by the structure product.

* Add xpsval to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is fairly
similar to the set.mm proof.

* Fix qus comments

This is qusaddval , qusaddf , qusmulval , and qusmulf .  Replace copy
pasted comments with ones which describe the theorems.  This is in
set.mm and iset.mm.